make example_tests a function and rename module
